requirements.txt
numpy==1.21.6
opencv_python==4.7.0.72
Pillow==9.5.0
Pillow==9.5.0
pylabel==0.1.49
ultralytics==8.0.92
